Renewal

+++ src/com/ephemeral/ktraytrace/IShape.kt
@@ -1,5 +1,5 @@
 package com.ephemeral.ktraytrace
 
 interface IShape {
-    fun intersect(ray: Ray, near: Float, far: Float) : Hit?
+    fun intersect(ray : Ray, near : Double, far : Double) : Hit?
 }
